LE_Kotlin: moved CoordinatesXY to a separate file, added OneAxis
